added ability to pass settings in URL
