+++ src/lib/components/settings/server-settings.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { persistentAppSettings } from '$lib/components/common/state.svelte';
-	import { testAPIConnectivity } from './server-settings.svelte.ts';
+	import { testAPIConnectivity } from './server-settings-functions.svelte.ts';
 	import { appSettings } from '$lib/components/common/state.svelte';
 	import { fly } from 'svelte/transition';
 
@@ -18,8 +18,12 @@
 		</label>
 		<label class="mb-2 block font-bold text-secondary" for="headscaleKey"> Headscale API Key </label>
 		<div class="relative flex">
-			<input id="headscaleKey" bind:value={persistentAppSettings.headscaleAPIKey} class="input input-sm input-bordered w-full" minlength="40" maxlength="40" type={apiSecretHidden ? "password" : "text"} required placeholder="******************" />
-			<button class="ml-2" onclick={() => {apiSecretHidden = !apiSecretHidden }}
+			<input id="headscaleKey" bind:value={persistentAppSettings.headscaleAPIKey} class="input input-sm input-bordered w-full" minlength="40" maxlength="40" type={apiSecretHidden ? 'password' : 'text'} required placeholder="******************" />
+			<button type="button"
+				class="ml-2"
+				onclick={() => {
+					apiSecretHidden = !apiSecretHidden;
+				}}
 				>{#if apiSecretHidden}
 					<!-- eye off -->
 					<svg xmlns="http://www.w3.org/2000/svg" class="my-1.5 h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="2"><path stroke-linecap="round" stroke-linejoin="round" d="M13.875 18.825A10.05 10.05 0 0112 19c-4.478 0-8.268-2.943-9.543-7a9.97 9.97 0 011.563-3.029m5.858.908a3 3 0 114.243 4.243M9.878 9.878l4.242 4.242M9.88 9.88l-3.29-3.29m7.532 7.532l3.29 3.29M3 3l3.59 3.59m0 0A9.953 9.953 0 0112 5c4.478 0 8.268 2.943 9.543 7a10.025 10.025 0 01-4.132 5.411m0 0L21 21" /></svg>
